feat: Add css modules and refactor Search components
