+++ src/components/search/Suggestions/index.ts
@@ -1,4 +1,2 @@
-export { default as SuggestionsTopSearch } from './SuggestionsTopSearch'
-export type { SuggestionsTopSearchProps } from './SuggestionsTopSearch'
-
 export { default } from './Suggestions'
+export type { SuggestionsProps } from './Suggestions'
