+++ src/components/search/SuggestionsTopSearch/SuggestionsTopSearch.tsx
@@ -8,6 +8,8 @@ import useSearchInput, { formatSearchPath } from 'src/sdk/search/useSearchInput'
 import useTopSearch from 'src/sdk/search/useTopSearch'
 import type { StoreSuggestionTerm } from '@generated/graphql'
 
+import styles from '../search-common.module.scss'
+
 export interface SuggestionsTopSearchProps
   extends HTMLAttributes<HTMLDivElement> {
   /**
@@ -30,10 +32,6 @@ const SuggestionsTopSearch = forwardRef<
   const { onSearchInputSelection } = useSearchInput()
   const { terms, isLoading } = useTopSearch(topTerms)
 
-  if (isLoading) {
-    return <p data-fs-search-input-loading-text>Loading...</p>
-  }
-
   if (terms.length === 0) {
     return null
   }
@@ -43,30 +41,37 @@ const SuggestionsTopSearch = forwardRef<
       ref={ref}
       data-testid={testId}
       data-fs-search-suggestion-section
+      className={styles.fsSearch}
       {...otherProps}
     >
-      <div data-fs-search-suggestion-header>
-        <p data-fs-search-suggestion-title>Top Search</p>
-      </div>
-      <UIList variant="ordered">
-        {terms.map((term, index) => (
-          <li key={term.value} data-fs-search-suggestion-item>
-            <Link
-              variant="display"
-              href={formatSearchPath(term.value)}
-              onClick={() =>
-                onSearchInputSelection?.(
-                  term.value,
-                  formatSearchPath(term.value)
-                )
-              }
-            >
-              <Badge variant="info">{index + 1}</Badge>
-              {term.value}
-            </Link>
-          </li>
-        ))}
-      </UIList>
+      {isLoading ? (
+        <p data-fs-search-input-loading-text>Loading...</p>
+      ) : (
+        <>
+          <div data-fs-search-suggestion-header>
+            <p data-fs-search-suggestion-title>Top Search</p>
+          </div>
+          <UIList variant="ordered">
+            {terms.map((term, index) => (
+              <li key={term.value} data-fs-search-suggestion-item>
+                <Link
+                  variant="display"
+                  href={formatSearchPath(term.value)}
+                  onClick={() =>
+                    onSearchInputSelection?.(
+                      term.value,
+                      formatSearchPath(term.value)
+                    )
+                  }
+                >
+                  <Badge variant="info">{index + 1}</Badge>
+                  {term.value}
+                </Link>
+              </li>
+            ))}
+          </UIList>
+        </>
+      )}
     </section>
   )
 })
