+++ src/components/search/SuggestionsTopSearch/index.ts
@@ -1,0 +1,2 @@
+export { default as SuggestionsTopSearch } from './SuggestionsTopSearch'
+export type { SuggestionsTopSearchProps } from './SuggestionsTopSearch'
