+++ src/components/search/search-common.module.scss
@@ -1,0 +1,74 @@
+@import "src/styles/scaffold";
+
+.fs-search {
+  [data-fs-search-suggestion-header] {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    min-height: var(--fs-spacing-7);
+  }
+
+  [data-fs-search-suggestion-title] {
+    font-size: var(--fs-text-size-3);
+    line-height: 1.5;
+  }
+
+  [data-fs-search-suggestion-title="small"] {
+    margin-bottom: var(--fs-spacing-1);
+    font-size: var(--fs-text-size-2);
+    color: var(--fs-color-neutral-6);
+  }
+
+  [data-fs-search-suggestion-section] {
+    padding: var(--fs-spacing-1) 0;
+    background-color: var(--fs-color-neutral-0);
+
+    &:not(:last-child) {
+      border-bottom: var(--fs-border-width) solid var(--fs-border-color-light);
+    }
+  }
+
+  // TODO: Temporarily increased CSS specificity because the normal link blue color
+  // was winning. Hoping this will be fixed when using CSS modules.
+  [data-fs-search-suggestions] [data-fs-search-suggestion-section="terms"] {
+    [data-fs-link] {
+      color: var(--fs-color-text-light);
+    }
+  }
+
+  [data-fs-search-suggestion-item] {
+    margin-right: calc(-1 * var(--fs-spacing-3));
+    margin-left: calc(-1 * var(--fs-spacing-3));
+    font-size: var(--fs-text-size-2);
+    line-height: 1.25;
+
+    &:hover {
+      background-color: var(--fs-color-main-0);
+    }
+
+    [data-fs-link] {
+      display: flex;
+      align-items: center;
+      justify-content: flex-start;
+      width: 100%;
+      padding-right: var(--fs-spacing-3);
+      padding-left: var(--fs-spacing-3);
+    }
+
+    [data-store-badge],
+    [data-fs-search-suggestion-item-icon] {
+      margin-right: var(--fs-spacing-1);
+    }
+
+    [data-fs-search-suggestion-item-icon] {
+      color: var(--fs-color-neutral-4);
+    }
+  }
+
+  [data-fs-search-suggestions] {
+    [data-fs-search-suggestion-item-bold] {
+      font-weight: var(--fs-text-weight-bold);
+      color: var(--fs-color-neutral-7);
+    }
+  }
+}
