--- src/components/ui/Search/Suggestions/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Suggestions'
-export type { SuggestionsProps } from './Suggestions'
