+++ src/styles/global/storybook-components.scss
@@ -5,9 +5,6 @@
 // Common
 @import "src/components/common/SearchInput/search-input.scss";
 
-// Search
-@import "src/components/search/Suggestions/suggestions.scss";
-
 // Sections
 @import "src/components/sections/BannerText/banner-text.scss";
 @import "src/components/sections/Breadcrumb/breadcrumb.scss";
